client: Move PKI methods to their own file
Just small cleanup to help make the module easier to navigate
